fix: 优化状态更新
