+++ task/buy.py
@@ -1,5 +1,4 @@
 import json
-import math
 import os
 import subprocess
 import sys
@@ -7,9 +6,7 @@ import time
 import uuid
 import copy
 import webbrowser
-import datetime
 from collections.abc import Generator
-from dataclasses import dataclass, field
 from json import JSONDecodeError
 import shutil
 import qrcode
@@ -17,375 +14,49 @@ from loguru import logger
 
 from requests import HTTPError, RequestException
 from cptoken import (
-    CTokenRuntimeState,
     generate_browser_window_state,
     init_ctoken_state,
-    sim_ctoken_state,
 )
 
-from util import time_service
 from util.Notifier import NotifierManager, NotifierConfig
 from util.ProxyBackoff import ProxyBackoff
 from util.BiliRequest import BiliRequest
 from util.ProxyManager import ProxyManager
 from util.RandomMessages import get_random_fail_message
 from util.TimeUtil import current_time_ms
-from util.TokenUtil import generate_token
 from util.error_codes import ErrorCodes
 from interface.project import fetch_project_payload
-
-
-base_url = "https://show.bilibili.com"
-BEIJING_TZ = datetime.timezone(datetime.timedelta(hours=8), name="Asia/Shanghai")
-WARMUP_AT_SECONDS = 5.0
-COUNTDOWN_REPORT_INTERVAL_SECONDS = 15
-
-
-@dataclass
-class BuyStreamState:
-    stage: str = "初始化"
-    countdown: str = "-"
-    countdown_seconds: int | None = None
-    current_proxy: str = "未初始化"
-    proxy_pool: str = ""
-    cooldown_remaining: int | None = None
-    attempt_current: int | None = None
-    attempt_total: int | None = None
-    payment_qr_url: str | None = None
-    status: str = "running"
-    last_message: str = ""
-
-
-@dataclass
-class BuyStreamEvent:
-    kind: str
-    message: str | None
-    state: BuyStreamState
-    data: dict = field(default_factory=dict)
-
-
-@dataclass
-class RetryOutcome:
-    err: int | None = None
-    ret: dict | None = None
-    exc: Exception | None = None
-
-    def set_response(self, err: int, ret: dict) -> None:
-        self.err = err
-        self.ret = ret
-        self.exc = None
-
-    def set_exception(self, exc: Exception) -> None:
-        self.exc = exc
-
-
-@dataclass(frozen=True)
-class CreateOrderTerminalRule:
-    status: str
-    message: str
-    expose_payment_url: bool = False
-
-
-def get_qrcode_url(_request, order_id) -> str:
-    url = f"{base_url}/api/ticket/order/getPayParam?order_id={order_id}"
-    data = _request.get(url).json()
-    if data.get("errno", data.get("code")) == 0:
-        return data["data"]["code_url"]
-    raise ValueError("获取二维码失败")
-
-
-def get_order_detail_url(order_id: int | str) -> str:
-    return f"{base_url}/platform/orderDetail.html?order_id={order_id}"
-
-
-def _format_countdown(seconds: float) -> str:
-    total_seconds = max(0, int(seconds))
-    hours, remainder = divmod(total_seconds, 3600)
-    minutes, secs = divmod(remainder, 60)
-    return f"{hours}小时{minutes}分{secs}秒"
-
-
-def _wait_until_start(time_start: str, warmup=None):
-    if not time_start:
-        return
-
-    timeoffset = time_service.get_timeoffset()
-    yield {"message": "0) 等待开始时间"}
-    yield {"message": f"时间偏差已被设置为: {timeoffset}秒"}
-
-    for fmt in (
-        "%Y-%m-%dT%H:%M:%S",
-        "%Y-%m-%d %H:%M:%S",
-        "%Y-%m-%dT%H:%M",
-        "%Y-%m-%d %H:%M",
-    ):
-        try:
-            target_time = datetime.datetime.strptime(time_start.strip(), fmt).replace(
-                tzinfo=BEIJING_TZ
-            )
-            break
-        except ValueError:
-            continue
-    else:
-        raise ValueError(f"无法解析抢票时间: {time_start!r}")
-
-    yield {"message": f"计划抢票开始时间: {target_time.strftime('%Y-%m-%d %H:%M:%S')}"}
-
-    time_difference = target_time.timestamp() - time.time() + timeoffset
-    end_time = time.perf_counter() + time_difference
-    next_report_at = float("inf")
-    warmed = False
-    last_countdown_seconds: int | None = None
-    while True:
-        remaining = end_time - time.perf_counter()
-        if remaining <= 0:
-            return
-        countdown_seconds = max(0, math.ceil(remaining))
-        countdown_text = _format_countdown(remaining)
-        if countdown_seconds != last_countdown_seconds:
-            last_countdown_seconds = countdown_seconds
-            yield {
-                "message": None,
-                "countdown": countdown_text,
-                "countdown_seconds": countdown_seconds,
-            }
-        if not warmed and warmup is not None and remaining <= WARMUP_AT_SECONDS:
-            warmed = True
-            for warm_message in warmup() or []:
-                yield {
-                    "message": warm_message,
-                    "countdown": countdown_text,
-                    "countdown_seconds": countdown_seconds,
-                }
-            continue
-        if countdown_seconds <= next_report_at:
-            yield {
-                "message": f"距离开始抢票还有: {countdown_text}",
-                "countdown": countdown_text,
-                "countdown_seconds": countdown_seconds,
-            }
-            next_report_at = max(
-                0, countdown_seconds - COUNTDOWN_REPORT_INTERVAL_SECONDS
-            )
-        time.sleep(min(0.5, remaining))
-
-
-def _build_token_payload(tickets_info: dict) -> dict:
-    count = int(tickets_info["count"])
-    screen_id = int(tickets_info["screen_id"])
-    order_type = int(tickets_info.get("order_type", 1))
-    project_id = int(tickets_info["project_id"])
-    sku_id = int(tickets_info["sku_id"])
-    return {
-        "count": count,
-        "screen_id": screen_id,
-        "order_type": order_type,
-        "project_id": project_id,
-        "sku_id": sku_id,
-        "buyer_info": tickets_info.get(
-            "_prepare_buyer_info",
-            tickets_info.get("buyer_info", []),
-        ),
-        "ignoreRequestLimit": True,
-        "ticket_agent": "",
-        "token": "",
-        "newRisk": True,
-        "requestSource": "neul-next",
-    }
-
-
-def _build_order_token(tickets_info: dict) -> str:
-    return generate_token(
-        project_id=int(tickets_info["project_id"]),
-        screen_id=int(tickets_info["screen_id"]),
-        order_type=int(tickets_info.get("order_type", 1)),
-        count=int(tickets_info["count"]),
-        sku_id=int(tickets_info["sku_id"]),
-    )
-
-
-def _normalize_prepare_ptoken(value: str | None) -> str:
-    if value is None:
-        return ""
-    return str(value).replace("=", "")
-
-
-CREATE_ORDER_TERMINAL_RULES: dict[int, CreateOrderTerminalRule] = {
-    100003: CreateOrderTerminalRule(
-        status="completed",
-        message="该项目每人限购1张，已存在购买订单，停止重试",
-    ),
-    100048: CreateOrderTerminalRule(
-        status="completed",
-        message="有尚未完成订单，停止重试",
-        expose_payment_url=True,
-    ),
-    100079: CreateOrderTerminalRule(
-        status="completed",
-        message="有重复订单，停止重试",
-    ),
-}
-
-
-def _create_order_terminal_rule(err: int) -> CreateOrderTerminalRule | None:
-    return CREATE_ORDER_TERMINAL_RULES.get(err)
-
-
-def _is_create_success(ret: dict, err: int) -> bool:
-    resp_message = str(ret.get("msg", ret.get("message", "")) or "")
-    return err == 0 and "defaultBBR" not in resp_message
-
-
-def _extract_order_id(ret: dict | None) -> int | str | None:
-    if not isinstance(ret, dict):
-        return None
-    data = ret.get("data")
-    if not isinstance(data, dict):
-        return None
-    order_id = data.get("orderId")
-    return order_id if order_id not in (None, "", 0) else None
-
-
-DEFAULT_CREATE_RETRY_LIMIT = 20
-DEFAULT_CREATE_REQUEST_BATCH_SIZE = 3
-DEFAULT_OUTER_LOOP_INTERVAL = 0
-
-
-def _extract_response_message(ret: dict) -> str:
-    return str(ret.get("msg", ret.get("message", "")) or "").strip()
-
-
-def _append_response_message(err: int, base: str, ret: dict | None) -> str:
-    return ErrorCodes.append_response_message(err, base, ret)
-
-
-def _format_retry_reason(outcome: RetryOutcome) -> str:
-    if outcome.exc is not None:
-        return f"最后一次异常: {outcome.exc}"
-    if outcome.err is None:
-        return "最后一次失败原因未知"
-    reason = ErrorCodes.get_message_or_unknown(outcome.err)
-    detail = outcome.ret if outcome.ret is not None else {}
-    base = f"最后一次返回: [{outcome.err}]({reason}) | {detail}"
-    return _append_response_message(outcome.err, base, outcome.ret)
-
-
-def _summarize_non_json_response(prefix: str, diagnostic: str) -> str:
-    if "status=412" in diagnostic:
-        return f"{prefix}触发 412 风控"
-
-    content_type = "未知"
-    for part in diagnostic.split(", "):
-        if part.startswith("content_type="):
-            content_type = part.split("=", 1)[1]
-            break
-    return f"{prefix}返回了非 JSON 响应（{content_type}）"
-
-
-def _build_proxy_exhausted_message(_request: BiliRequest, delay_seconds: int) -> str:
-    return (
-        "当前所有代理暂时不可用，请尽快补充或更换代理。"
-        f"程序将休息 {delay_seconds} 秒后继续尝试。"
-        f" 代理池状态：{_request.proxy_pool_status()}"
-    )
-
-
-def _notify_proxy_exhausted(
-    notifier_config: NotifierConfig,
-    _request: BiliRequest,
-    delay_seconds: int,
-) -> None:
-    if not notifier_config.notify_proxy_exhausted:
-        return
-
-    manager = NotifierManager.create_from_config(
-        config=notifier_config,
-        title="代理已全部失效",
-        content=_build_proxy_exhausted_message(_request, delay_seconds),
-        include_audio=False,
-    )
-    manager.start_all()
-
-
-def _handle_proxy_failure(
-    _request: BiliRequest,
-    reason: str,
-    proxy_backoff: ProxyBackoff,
-    notifier_config: NotifierConfig,
-) -> tuple[str | None, int | None]:
-    """Handle a proxy failure and return the immediate status plus cooldown plan."""
-    previous_proxy = _request.current_proxy_display()
-    cooled = _request.mark_current_proxy_failure(reason)
-    if cooled:
-        immediate_message = f"代理冷却: {previous_proxy} 短时间内连续失败，已暂时停用"
-    else:
-        immediate_message = None
-
-    if _request.switch_proxy():
-        proxy_backoff.reset()
-        switched_message = f"切换代理到 {_request.current_proxy_display()}"
-        if immediate_message:
-            return f"{immediate_message}\n{switched_message}", None
-        return switched_message, None
-
-    if _request.has_available_proxy():
-        return immediate_message, None
-
-    delay_seconds = proxy_backoff.next_delay_seconds()
-    if proxy_backoff.should_notify():
-        _notify_proxy_exhausted(notifier_config, _request, delay_seconds)
-    exhausted_message = f"所有代理当前不可用，休息 {delay_seconds} 秒后再试"
-    if immediate_message:
-        return f"{immediate_message}\n{exhausted_message}", delay_seconds
-    return exhausted_message, delay_seconds
-
-
-def _format_status_result(prefix: str, ret: dict) -> str:
-    err = int(ret.get("errno", ret.get("code", -1)))
-    reason = ErrorCodes.get_message(err)
-    if reason:
-        return _append_response_message(err, f"{prefix}: [{err}] {reason}", ret)
-    message = _extract_response_message(ret)
-    if message:
-        return f"{prefix}: [{err}] {message}"
-    return f"{prefix}: [{err}] {ret}"
-
-
-def _prepare_create_request(
-    tickets_info: dict,
-    order_token: str,
-    is_hot_project: bool,
-    request_result: dict | None,
-    ticket_state: CTokenRuntimeState,
-) -> tuple[str, dict]:
-    payload = dict(tickets_info)
-    payload["again"] = 1
-    payload["token"] = order_token
-    now_ms = current_time_ms()
-    payload["timestamp"] = now_ms
-    payload["newRisk"] = True
-    payload["requestSource"] = "neul-next"
-    payload.pop("detail", None)
-    payload.pop("sale_start", None)
-    payload.pop("username", None)
-    payload.pop("_prepare_buyer_info", None)
-    url = (
-        f"{base_url}/api/ticket/order/createV2?project_id={tickets_info['project_id']}"
-    )
-
-    if not is_hot_project:
-        return url, payload
-    create_state = sim_ctoken_state(
-        before_state=ticket_state,
-        now_ms=now_ms,
-    )
-    payload["ctoken"] = create_state.generate_create_ctoken()
-    prepare_data = request_result.get("data", {}) if request_result else {}
-    ptoken = _normalize_prepare_ptoken(prepare_data.get("ptoken"))
-    payload["ptoken"] = ptoken
-    payload["orderCreateUrl"] = "https://show.bilibili.com/api/ticket/order/createV2"
-    url += "&ptoken=" + ptoken
-    return url, payload
+from task.buy_helpers import (
+    BASE_URL as base_url,
+    DEFAULT_CREATE_REQUEST_BATCH_SIZE,
+    DEFAULT_CREATE_RETRY_LIMIT,
+    DEFAULT_OUTER_LOOP_INTERVAL,
+    build_order_token as _build_order_token,
+    build_token_payload as _build_token_payload,
+    create_order_terminal_rule as _create_order_terminal_rule,
+    extract_order_id as _extract_order_id,
+    format_retry_reason as _format_retry_reason,
+    format_status_result as _format_status_result,
+    get_order_detail_url,
+    get_qrcode_url,
+    handle_proxy_failure as _handle_proxy_failure,
+    is_create_success as _is_create_success,
+    prepare_create_request as _prepare_create_request,
+    summarize_non_json_response as _summarize_non_json_response,
+    wait_until_start as _wait_until_start,
+)
+from task.buy_types import (
+    BuyStreamEvent,
+    BuyStreamState,
+    BuyStreamUpdate,
+    BuyStreamWorker,
+    CreateOrderTerminalRule,
+    RetryOutcome,
+)
+
+
+def start_buy_stream_worker(*args, **kwargs) -> BuyStreamWorker:
+    return BuyStreamWorker(buy_stream, *args, **kwargs).start()
 
 
 def buy_stream(
@@ -407,27 +78,13 @@ def buy_stream(
 ):
     state = BuyStreamState()
 
-    def emit(kind: str, message: str | None, **data):
-        if "stage" in data:
-            state.stage = data["stage"]
-        if "countdown" in data:
-            state.countdown = data["countdown"]
-        if "countdown_seconds" in data:
-            state.countdown_seconds = data["countdown_seconds"]
-        if "current_proxy" in data:
-            state.current_proxy = data["current_proxy"]
-        if "proxy_pool" in data:
-            state.proxy_pool = data["proxy_pool"]
-        if "cooldown_remaining" in data:
-            state.cooldown_remaining = data["cooldown_remaining"]
-        if "attempt_current" in data:
-            state.attempt_current = data["attempt_current"]
-        if "attempt_total" in data:
-            state.attempt_total = data["attempt_total"]
-        if "payment_qr_url" in data:
-            state.payment_qr_url = data["payment_qr_url"]
-        if "status" in data:
-            state.status = data["status"]
+    def emit(
+        kind: str,
+        message: str | None,
+        update: BuyStreamUpdate | None = None,
+    ):
+        if update is not None:
+            update.apply_to(state)
         if message is not None:
             state.last_message = message
 
@@ -435,7 +92,7 @@ def buy_stream(
             kind=kind,
             message=message,
             state=copy.deepcopy(state),
-            data=data,
+            data=update.to_dict() if update is not None else {},
         )
 
     def handle_proxy_failure(
@@ -457,12 +114,14 @@ def buy_stream(
                 yield emit(
                     "proxy",
                     message,
-                    current_proxy=_request.current_proxy_status(),
-                    proxy_pool=_request.proxy_pool_status(),
-                    cooldown_remaining=None,
-                    status="running",
-                    attempt_current=attempt,
-                    attempt_total=attempt_total,
+                    BuyStreamUpdate(
+                        current_proxy=_request.current_proxy_status(),
+                        proxy_pool=_request.proxy_pool_status(),
+                        cooldown_remaining=None,
+                        status="running",
+                        attempt_current=attempt,
+                        attempt_total=attempt_total,
+                    ),
                 )
         if delay_seconds is None:
             return
@@ -470,12 +129,14 @@ def buy_stream(
             yield emit(
                 "state",
                 None,
-                current_proxy=_request.current_proxy_status(),
-                proxy_pool=_request.proxy_pool_status(),
-                cooldown_remaining=remaining,
-                status="cooldown",
-                attempt_current=attempt,
-                attempt_total=attempt_total,
+                BuyStreamUpdate(
+                    current_proxy=_request.current_proxy_status(),
+                    proxy_pool=_request.proxy_pool_status(),
+                    cooldown_remaining=remaining,
+                    status="cooldown",
+                    attempt_current=attempt,
+                    attempt_total=attempt_total,
+                ),
             )
             time.sleep(1)
         if _request.ensure_active_proxy():
@@ -483,12 +144,14 @@ def buy_stream(
             yield emit(
                 "state",
                 None,
-                current_proxy=_request.current_proxy_status(),
-                proxy_pool=_request.proxy_pool_status(),
-                cooldown_remaining=None,
-                status="running",
-                attempt_current=attempt,
-                attempt_total=attempt_total,
+                BuyStreamUpdate(
+                    current_proxy=_request.current_proxy_status(),
+                    proxy_pool=_request.proxy_pool_status(),
+                    cooldown_remaining=None,
+                    status="running",
+                    attempt_current=attempt,
+                    attempt_total=attempt_total,
+                ),
             )
 
     def handle_non_json_response(
@@ -504,6 +167,23 @@ def buy_stream(
             yield emit(
                 "proxy",
                 f"{prefix}触发 412 风控",
+                BuyStreamUpdate(
+                    current_proxy=_request.current_proxy_status(),
+                    proxy_pool=_request.proxy_pool_status(),
+                    attempt_current=attempt,
+                    attempt_total=(
+                        effective_retry_limit
+                        if attempt is not None
+                        else state.attempt_total
+                    ),
+                ),
+            )
+            yield from handle_proxy_failure(f"{prefix} 412 风控", attempt=attempt)
+            return True
+        yield emit(
+            "attempt" if attempt is not None else "error",
+            summary,
+            BuyStreamUpdate(
                 current_proxy=_request.current_proxy_status(),
                 proxy_pool=_request.proxy_pool_status(),
                 attempt_current=attempt,
@@ -512,17 +192,6 @@ def buy_stream(
                     if attempt is not None
                     else state.attempt_total
                 ),
-            )
-            yield from handle_proxy_failure(f"{prefix} 412 风控", attempt=attempt)
-            return True
-        yield emit(
-            "attempt" if attempt is not None else "error",
-            summary,
-            current_proxy=_request.current_proxy_status(),
-            proxy_pool=_request.proxy_pool_status(),
-            attempt_current=attempt,
-            attempt_total=(
-                effective_retry_limit if attempt is not None else state.attempt_total
             ),
         )
         return False
@@ -578,19 +247,23 @@ def buy_stream(
         yield emit(
             "status",
             wait_message,
-            stage=stage_value or state.stage,
-            countdown=countdown_value or state.countdown,
-            countdown_seconds=(
-                countdown_seconds
-                if countdown_seconds is not None
-                else state.countdown_seconds
+            BuyStreamUpdate(
+                stage=stage_value or state.stage,
+                countdown=countdown_value or state.countdown,
+                countdown_seconds=(
+                    countdown_seconds
+                    if countdown_seconds is not None
+                    else state.countdown_seconds
+                ),
             ),
         )
     yield emit(
         "proxy",
         f"当前代理: {_request.current_proxy_status()}",
-        current_proxy=_request.current_proxy_status(),
-        proxy_pool=_request.proxy_pool_status(),
+        BuyStreamUpdate(
+            current_proxy=_request.current_proxy_status(),
+            proxy_pool=_request.proxy_pool_status(),
+        ),
     )
 
     while isRunning:
@@ -607,7 +280,7 @@ def buy_stream(
             )
             if is_hot_project:
                 # hot
-                yield emit("stage", "开始准备订单", stage="订单准备")
+                yield emit("stage", "开始准备订单", BuyStreamUpdate(stage="订单准备"))
                 prepare_ctoken_state = ticket_state.snapshot(now_ms=ticket_collection_t)
                 token_payload["token"] = prepare_ctoken_state.generate_prepare_ctoken()
                 request_result_normal = _request.post(
@@ -629,7 +302,7 @@ def buy_stream(
 
             else:
                 # normal
-                yield emit("status", None, stage="订单准备")
+                yield emit("status", None, BuyStreamUpdate(stage="订单准备"))
                 if use_local_token:
                     order_token = _build_order_token(tickets_info)
                     yield emit(
@@ -653,9 +326,11 @@ def buy_stream(
             yield emit(
                 "stage",
                 "开始创建订单",
-                stage="创建订单",
-                attempt_current=None,
-                attempt_total=effective_retry_limit,
+                BuyStreamUpdate(
+                    stage="创建订单",
+                    attempt_current=None,
+                    attempt_total=effective_retry_limit,
+                ),
             )
             result = None
             retry_outcome = RetryOutcome()
@@ -676,7 +351,7 @@ def buy_stream(
                 )
                 while attempt <= batch_end:
                     if not isRunning:
-                        yield "抢票结束"
+                        yield emit("status", "抢票结束")
                         break
                     try:
                         create_response = _request.post(
@@ -692,11 +367,21 @@ def buy_stream(
                             yield emit(
                                 "success",
                                 "创建订单成功",
-                                attempt_current=attempt,
-                                attempt_total=effective_retry_limit,
+                                BuyStreamUpdate(
+                                    attempt_current=attempt,
+                                    attempt_total=effective_retry_limit,
+                                ),
                             )
                             result = (ret, err)
                             break
+                        yield emit(
+                            "attempt",
+                            ErrorCodes.format_attempt_result(err, ret),
+                            BuyStreamUpdate(
+                                attempt_current=attempt,
+                                attempt_total=effective_retry_limit,
+                            ),
+                        )
                         terminal_rule = _create_order_terminal_rule(err)
                         if terminal_rule is not None:
                             terminal_result = (err, ret, terminal_rule)
@@ -707,27 +392,25 @@ def buy_stream(
                                     terminal_rule.message,
                                     ret,
                                 ),
-                                attempt_current=attempt,
-                                attempt_total=effective_retry_limit,
-                                status=terminal_rule.status,
+                                BuyStreamUpdate(
+                                    attempt_current=attempt,
+                                    attempt_total=effective_retry_limit,
+                                    status=terminal_rule.status,
+                                ),
                             )
                             break
                         if err == 100051:
                             yield emit("status", "token过期，需要重新准备订单")
                             token_expired = True
                             break
-                        yield emit(
-                            "attempt",
-                            ErrorCodes.format_attempt_result(err, ret),
-                            attempt_current=attempt,
-                            attempt_total=effective_retry_limit,
-                        )
                         if err == 100034:
                             yield emit(
                                 "status",
                                 f"更新票价为：{ret['data']['pay_money'] / 100}",
-                                attempt_current=attempt,
-                                attempt_total=effective_retry_limit,
+                                BuyStreamUpdate(
+                                    attempt_current=attempt,
+                                    attempt_total=effective_retry_limit,
+                                ),
                             )
                             tickets_info["pay_money"] = ret["data"]["pay_money"]
                         time.sleep(inner_loop_interval / 1000)
@@ -751,16 +434,21 @@ def buy_stream(
                         yield emit(
                             "attempt",
                             str(e),
-                            attempt_current=attempt,
-                            attempt_total=effective_retry_limit,
+                            BuyStreamUpdate(
+                                attempt_current=attempt,
+                                attempt_total=effective_retry_limit,
+                            ),
                         )
                     except Exception as e:
+                        logger.exception(e)
                         retry_outcome.set_exception(e)
                         yield emit(
                             "attempt",
                             str(e),
-                            attempt_current=attempt,
-                            attempt_total=effective_retry_limit,
+                            BuyStreamUpdate(
+                                attempt_current=attempt,
+                                attempt_total=effective_retry_limit,
+                            ),
                         )
                     if (
                         result is not None
@@ -787,8 +475,9 @@ def buy_stream(
                 yield emit(
                     "status",
                     None,
-                    attempt_current=None,
-                    attempt_total=effective_retry_limit,
+                    BuyStreamUpdate(
+                        attempt_total=effective_retry_limit,
+                    ),
                 )
                 continue
             if result is None:
@@ -800,8 +489,10 @@ def buy_stream(
                         yield emit(
                             "payment_qr",
                             "PAYMENT_QR_URL={0}".format(payment_url),
-                            payment_qr_url=payment_url,
-                            status=terminal_rule.status,
+                            BuyStreamUpdate(
+                                payment_qr_url=payment_url,
+                                status=terminal_rule.status,
+                            ),
                         )
                         if auto_open_payment_url:
                             try:
@@ -809,8 +500,10 @@ def buy_stream(
                                 yield emit(
                                     "status",
                                     "已自动打开现有订单链接",
-                                    payment_qr_url=payment_url,
-                                    status=terminal_rule.status,
+                                    BuyStreamUpdate(
+                                        payment_qr_url=payment_url,
+                                        status=terminal_rule.status,
+                                    ),
                                 )
                             except Exception as exc:
                                 yield emit("status", f"自动打开订单链接失败: {exc}")
@@ -835,8 +528,10 @@ def buy_stream(
                 yield emit(
                     "stage",
                     "抢票成功，弹出付款二维码",
-                    stage="抢票成功",
-                    status="succeeded",
+                    BuyStreamUpdate(
+                        stage="抢票成功",
+                        status="succeeded",
+                    ),
                 )
                 order_id = request_result["data"]["orderId"]  # type: ignore
                 payment_url = get_order_detail_url(order_id)
@@ -847,8 +542,10 @@ def buy_stream(
                 yield emit(
                     "payment_qr",
                     "PAYMENT_QR_URL={0}".format(payment_url),
-                    payment_qr_url=payment_url,
-                    status="succeeded",
+                    BuyStreamUpdate(
+                        payment_qr_url=payment_url,
+                        status="succeeded",
+                    ),
                 )
                 if auto_open_payment_url:
                     try:
@@ -856,8 +553,10 @@ def buy_stream(
                         yield emit(
                             "status",
                             "已自动打开支付链接",
-                            payment_qr_url=payment_url,
-                            status="succeeded",
+                            BuyStreamUpdate(
+                                payment_qr_url=payment_url,
+                                status="succeeded",
+                            ),
                         )
                     except Exception as exc:
                         yield emit("status", f"自动打开支付链接失败: {exc}")
@@ -877,7 +576,11 @@ def buy_stream(
                 yield message
         except Exception as e:
             logger.exception(e)
-            yield emit("error", f"程序异常: {repr(e)}", status="failed")
+            yield emit(
+                "error",
+                f"程序异常: {repr(e)}",
+                BuyStreamUpdate(status="failed"),
+            )
 
 
 def buy(
